backend/requirements.txt
Django>=5.2.10
djangorestframework>=3.16.1
django-cors-headers>=4.4.0
Pillow>=10.0.0
dj-database-url>=2.2.0
gunicorn>=23.0.0
psycopg[binary]>=3.2.3
whitenoise>=6.7.0
boto3>=1.34.0
django-storages>=1.14.0
